pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vyntri"
version = "0.1.0"
description = "Training-less vision intelligence for edge devices"
authors = [
  { name = "Vyntri Contributors" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "scikit-learn>=1.3.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
webcam = ["opencv-python>=4.8.0"]

[project.scripts]
vyntri = "vyntri.cli:main"

[project.urls]
"Homepage" = "https://github.com/vyntri/vyntri"
"Bug Tracker" = "https://github.com/vyntri/vyntri/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["vyntri*"]
